# Add perf test for time zone operators