closes [#10142](https://github.com/NVIDIA/spark-rapids/issues/10142)

Test result:
|operator|tz|mean cpu|mean gpu|ratio|
|:-|:-|:-|:-|:-|
|from_utc_timestamp|Asia/Shanghai |689.20 ms       | 111.00 ms |6.21 x|
|from_utc_timestamp|Japan |608.00 ms       | 92.60 ms |6.57 x|
|to_utc_timestamp|Asia/Shanghai|811.20 ms       | 84.80 ms |9.57 x|
|to_utc_timestamp| Japan|737.40 ms       | 82.60 ms |8.93 x|
|hour| Asia/Shanghai|565.80 ms       | 83.40 ms |6.78 x|
|hour|Japan|520.60 ms       | 77.80 ms |6.69 x|
|minute| Asia/Shanghai|546.80 ms       | 75.40 ms |7.25 x|
|minute|Japan|510.60 ms       | 73.40 ms |6.96 x|
|second| Asia/Shanghai |548.80 ms       | 74.20 ms |7.40 x|
|second|Japan|517.60 ms       | 70.40 ms |7.35 x|
|unix_timestamp| Asia/Shanghai |3439.00 ms       | 76.20 ms |45.13 x
|unix_timestamp |Japan|3410.20 ms       | 75.60 ms |45.11 x|
|from_unixtime | Asia/Shanghai  |3494.40 ms      | 81.20 ms |43.03 x|
|from_unixtime |Japan | 3309.80 ms      | 76.00 ms |43.55 x|
|date_format |Asia/Shanghai|3399.60 ms      | 72.60 ms |46.83 x|
|date_format | Japan  |3373.80 ms      | 73.40 ms |45.96 x|

Signed-off-by: Chong Gao <res_life@163.com>